requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.29
pydantic-settings>=2.2.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
cryptography>=42.0.5
apscheduler>=3.10.4
jinja2>=3.1.3
python-multipart>=0.0.9
twilio>=9.0.0
pytz>=2024.1
httpx>=0.27.0
pytest>=8.1.1
pytest-asyncio>=0.23.6
